Sonar fixes, somes codes improvements and others
Fixed a lot of sonar checks
Created new function "getTimeNow" in Game class, for get the real time using std::chrono
Created function "setPosition" for Position class
Removed functions "readAttr" from classes; Teleport, Bed, Container, DepotLocker, Door
The attribute is now read directly from item.cpp, and classes just return virtual functions

IOMap::loadMap
Changed BinaryNode for use "shared_ptr"
Various improvements on map loading, added several sanity checks and other improvements
Removed "setLastErrorString" function
Added instead just the spdlog log
Improved the function of checking teleports

Refactor DBResult function "getNumber"

The function "DBResult::getNumber has been broken into several smaller functions, instead of using a generalized function, we will use a function for each value, with sanity checks and using the new features of c++ to perform the string conversions to the appropriate types (int/uint, time_t/boolean).

With this change, we've prevented data loss and added some sanity checks to ensure that if there is a problem, we have a proper log of what happened.

New c++ functions
DBResult::get8()
DBResult::get16()
DBResult::get32()
DBResult::get64()

DBResult::getU8()
DBResult::getU16()
DBResult::getU32()
DBResult::getU64()

DBResult::getTime()
DBResult::getBoolean()

New Lua functions
result:get8()
result:get16()
result:get32()
result:get64()

result:getU8()
result:getU16()
result:getU32()
result:getU64()

result:getTime()
result:getBoolean()

Change negative comparison to call feature of c++20 "std::cmp_less"

Minimum required version for visual studio

Fixed remaining conflict resolution bugs

Sonar fixes, somes codes improvements and others

Fixed a lot of sonar checks
Created new function "getTimeNow" in Game class, for get the real time using std::chrono
Created function "setPosition" for Position class
Removed functions "readAttr" from classes; Teleport, Bed, Container, DepotLocker, Door
The attribute is now read directly from item.cpp, and classes just return virtual functions

IOMap::loadMap
Changed BinaryNode for use "shared_ptr"
Various improvements on map loading, added several sanity checks and other improvements
Removed "setLastErrorString" function
Added instead just the spdlog log
Improved the function of checking teleports

Refactor DBResult function "getNumber"

The function "DBResult::getNumber has been broken into several smaller functions, instead of using a generalized function, we will use a function for each value, with sanity checks and using the new features of c++ to perform the string conversions to the appropriate types (int/uint, time_t/boolean).

With this change, we've prevented data loss and added some sanity checks to ensure that if there is a problem, we have a proper log of what happened.

New c++ functions
DBResult::get8()
DBResult::get16()
DBResult::get32()
DBResult::get64()

DBResult::getU8()
DBResult::getU16()
DBResult::getU32()
DBResult::getU64()

DBResult::getTime()
DBResult::getBoolean()

New Lua functions
result:get8()
result:get16()
result:get32()
result:get64()

result:getU8()
result:getU16()
result:getU32()
result:getU64()

result:getTime()
result:getBoolean()

Change negative comparison to call feature of c++20 "std::cmp_less"

Minimum required version for visual studio

Fixed remaining conflict resolution bugs

Sonar fixes, somes codes improvements and others

Fixed a lot of sonar checks
Created new function "getTimeNow" in Game class, for get the real time using std::chrono
Created function "setPosition" for Position class
Removed functions "readAttr" from classes; Teleport, Bed, Container, DepotLocker, Door
The attribute is now read directly from item.cpp, and classes just return virtual functions

IOMap::loadMap
Changed BinaryNode for use "shared_ptr"
Various improvements on map loading, added several sanity checks and other improvements
Removed "setLastErrorString" function
Added instead just the spdlog log
Improved the function of checking teleports

Refactor DBResult function "getNumber"

The function "DBResult::getNumber has been broken into several smaller functions, instead of using a generalized function, we will use a function for each value, with sanity checks and using the new features of c++ to perform the string conversions to the appropriate types (int/uint, time_t/boolean).

With this change, we've prevented data loss and added some sanity checks to ensure that if there is a problem, we have a proper log of what happened.

New c++ functions
DBResult::get8()
DBResult::get16()
DBResult::get32()
DBResult::get64()

DBResult::getU8()
DBResult::getU16()
DBResult::getU32()
DBResult::getU64()

DBResult::getTime()
DBResult::getBoolean()

New Lua functions
result:get8()
result:get16()
result:get32()
result:get64()

result:getU8()
result:getU16()
result:getU32()
result:getU64()

result:getTime()
result:getBoolean()

Change negative comparison to call feature of c++20 "std::cmp_less"

Minimum required version for visual studio

Fixed remaining conflict resolution bugs

+++ src/lua/functions/creatures/player/player_functions.cpp
@@ -1239,7 +1239,7 @@ int PlayerFunctions::luaPlayerSetVocation(lua_State* L) {
 	if (isNumber(L, 2)) {
 		vocation = g_vocations().getVocation(getNumber<uint16_t>(L, 2));
 	} else if (isString(L, 2)) {
-		vocation = g_vocations().getVocation(g_vocations().getVocationId(getString(L, 2)));
+		vocation = g_vocations().getVocation(static_cast<uint16_t>(g_vocations().getVocationId(getString(L, 2))));
 	} else if (isUserdata(L, 2)) {
 		vocation = getUserdata<Vocation>(L, 2);
 	} else {
